refactor(viewer): delete unused useMeasurementPointHandler hook
