+++ libs/insight-viewer/src/Annotation/utils/getCircleEndPoint.ts
@@ -1,4 +1,4 @@
-import type { Point } from '../../../types'
+import type { Point } from '../../types'
 
 export function getCircleEndPoint(centerPoint: Point, radius: number): Point {
   const endPoint: Point = [centerPoint[0] + radius, centerPoint[1]]
