+++ libs/insight-viewer/src/Annotation/utils/getCircleStartPoint.ts
@@ -1,4 +1,4 @@
-import type { Point } from '../../../types'
+import type { Point } from '../../types'
 
 export function getCircleStartPoint(centerPoint: Point, radius: number): Point {
   const startPoint: Point = [centerPoint[0] - radius, centerPoint[1]]
