+++ libs/insight-viewer/src/Annotation/utils/getEditingTextPosition.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import { Point } from '../../types'
 
 export function getEditingTextPosition(currentPoint: Point): Point {
   return currentPoint
