+++ libs/insight-viewer/src/Annotation/utils/getLineLength.spec.ts
@@ -1,5 +1,5 @@
-import { Point } from '../../../types'
-import { Image } from '../../../Viewer/types'
+import { Point } from '../../types'
+import { Image } from '../../Viewer/types'
 import { getLineLength } from './getLineLength'
 
 describe('getLineLength: ', () => {
