--- libs/insight-viewer/src/hooks/useMeasurementPointsHandler/utils/getMeasurementDrawingPoints.ts
@@ -1,13 +1,0 @@
-import { Point, MeasurementMode } from '../../../types'
-
-export function getMeasurementDrawingPoints(
-  prevPoints: [Point, Point],
-  currentPoint: Point,
-  mode: MeasurementMode
-): [Point, Point] {
-  if (mode === 'ruler' || mode === 'area') {
-    return [prevPoints[0], currentPoint]
-  }
-
-  return prevPoints
-}
